+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+0.2.1 (2015-07-29)
+---------------------
+
+* Remove `Shipment.is_booked()` method. Use `Shipment.state` instead.
+* Raise `ShipmentNotCompletedException` from `Shipment.track()` method
+
 0.2.0 (2015-07-28)
 ---------------------